% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{BoardSize}
\alias{BoardSize}
\title{the number of columns in the tic-tac-toe board}
\usage{
BoardSize()
}
\value{
Integer
}
\description{
the number of columns in the tic-tac-toe board
}
\examples{
BoardSize()
}
