% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{DisableBoard}
\alias{DisableBoard}
\title{disables both checkboxes if cell has been played.
otherwise disable so one side (X,O) cannot play.}
\usage{
DisableBoard(localBoard, computerSide, xSide, oSide)
}
\arguments{
\item{localBoard:}{vector (1:9) - values for a cell: -1, 0, 1}

\item{computerSide:}{Integer: (1,-1)}

\item{xSide:}{Integer: (1,-1)}

\item{oSide:}{Integer: (1,-1)}
}
\value{

}
\description{
disables both checkboxes if cell has been played.
otherwise disable so one side (X,O) cannot play.
}
\examples{
DisableBoard()
}
