% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{EvaluateBoard}
\alias{EvaluateBoard}
\title{returns who has won if any one.}
\usage{
EvaluateBoard(board)
}
\arguments{
\item{board:}{3x3 matrix representing board - values for a cell: -1, 0, 1}
}
\value{
Integer: -1 or 1, or 0 if no-one has won.
}
\description{
returns who has won if any one.
}
\examples{
EvaluateBoard()
}
