% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{IsMovePossible}
\alias{IsMovePossible}
\title{returns whether are any moves left.}
\usage{
IsMovePossible(board)
}
\arguments{
\item{board:}{3x3 matrix representing board - values for a cell: -1, 0, 1}
}
\value{
Bollean: TRUE if some cells have 0 otherwise FALSE
}
\description{
returns whether are any moves left.
}
\examples{
EvaluateBoard()
}
