% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{UICellForIndex}
\alias{UICellForIndex}
\title{converts index on the board to a list representing the
X and O checkboxes for the cell.}
\usage{
UICellForIndex(index)
}
\arguments{
\item{index:}{integer 1:9}
}
\value{
list with checkbox names for cell.
}
\description{
converts index on the board to a list representing the
X and O checkboxes for the cell.
}
\examples{
UICellForIndex()
}
