% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{UpdateBoard}
\alias{UpdateBoard}
\title{shiny ui for a tic-tac-toe game}
\usage{
UpdateBoard(input, xSide = -999, oSide = 999)
}
\arguments{
\item{input}{xSide
oSide}
}
\description{
shiny ui for a tic-tac-toe game
}
\examples{
shinyTTTUI()
}
