% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBoard.R
\name{UpdateInputForMove}
\alias{UpdateInputForMove}
\title{updates UI side with serverside move}
\usage{
UpdateInputForMove(localBoard, move, side, xSide, session)
}
\arguments{
\item{localBoard:}{vector (1:9) - values for a cell: -1, 0, 1}

\item{move:}{Integer: (-1, 0, 1)}

\item{side:}{Integer: (-1, 1)}

\item{xSide:}{Integer: (-1, 1)}

\item{session:}{shiny session}
}
\value{
list with checkbox names for cell.
}
\description{
updates UI side with serverside move
}
\examples{
UpdateInputForMove()
}
